+++ p5-analysis/src/index.ts
@@ -1,7 +1,7 @@
 export { Category } from './lib/Category';
 export { Library } from './lib/Library';
 export { JavaScriptSyntaxError, Script } from './lib/Script';
-export { Sketch, SketchType } from './lib/Sketch';
+export { Sketch, SketchStructureType } from './lib/Sketch';
 
 import { Category } from './lib/Category';
 Category.load();