+++ p5-server/src/commands/build.ts
@@ -2,11 +2,11 @@ import fs from 'fs';
 import { rm as rmSync, writeFile } from 'fs/promises';
 import minimatch from 'minimatch';
 import open from 'open';
-import { Sketch } from 'p5-analysis';
+import { Sketch, SketchStructureType } from 'p5-analysis';
 import path from 'path';
 import {
   createDirectoryListing,
-  defaultDirectoryExclusions
+  defaultDirectoryExclusions,
 } from '../server/createDirectoryListing';
 import { markdownToHtmlPage, sourceViewTemplate } from '../server/templates';
 import { die, pathIsInDirectory, stringToOptions } from '../utils';
@@ -99,11 +99,13 @@ function createActions(file: string, output: string): ActionIterator {
 
   async function* visitDir(dir: string, output: string): ActionIterator {
     const { sketches, allFiles } = await Sketch.analyzeDirectory(dir, {
-      exclusions: directoryExclusions
+      exclusions: directoryExclusions,
     });
     yield Action('mkdir', dir, output);
 
-    const scriptOnlySketches = sketches.filter(sk => sk.sketchType === 'javascript');
+    const scriptOnlySketches = sketches.filter(
+      sk => sk.structureType === SketchStructureType.scriptOnly
+    );
     // TODO: check for collisions when choosing the output file path
     for (const sketch of scriptOnlySketches) {
       const outputFile = path
@@ -144,7 +146,7 @@ function createActions(file: string, output: string): ActionIterator {
         kind: 'createIndex',
         dir,
         outputFile,
-        path: path.basename(dir)
+        path: path.basename(dir),
       };
     }
   }
@@ -212,7 +214,7 @@ async function runActions(actions: ActionIterator, options: Options) {
         const html = await createDirectoryListing(dir, path, {
           staticMode: true,
           templateName: options.theme,
-          templateOptions
+          templateOptions,
         });
         await writeFile(outputFile, html);
         filesCreated += 1;