+++ p5-server/src/commands/convert.ts
@@ -1,21 +1,41 @@
-import { Sketch, SketchType } from 'p5-analysis';
+import { Sketch, SketchStructureType } from 'p5-analysis';
 import { die } from '../utils';
+import fs from 'fs';
+import path from 'path';
 
-const sketchTypes: SketchType[] = ['html', 'javascript'];
+const sketchTypes: Record<string, SketchStructureType | 'folder'> = {
+  '^html$': SketchStructureType.htmlIndex,
+  '^(script|javascript|js)(-only)?$': SketchStructureType.scriptOnly,
+  '^folder$': 'folder',
+};
 
-export default async function convert(
-  sketchPath: string,
-  options: { to?: SketchType }
-) {
+export default async function convert(sketchPath: string, options: { to: string }) {
+  if (!options.to) {
+    die(`Missing required option: --to`);
+  }
+  const targetType: SketchStructureType | 'folder' =
+    Object.entries(sketchTypes).find(([regex]) => options.to.match(regex))?.[1] ??
+    die(`Invalid option --to ${options.to}`);
+
+  if (!fs.existsSync(sketchPath) && fs.existsSync(sketchPath + '.html'))
+    sketchPath += '.html';
+  if (!fs.existsSync(sketchPath) && fs.existsSync(sketchPath + '.js'))
+    sketchPath += '.js';
+  // TODO: if it's a script file that belongs to an HTML index in the same directory, warn or rename the index instead
   const sketch = await Sketch.fromFile(sketchPath);
-  const targetType =
-    options.to || (sketch.sketchType === 'html' ? 'javascript' : 'html');
 
-  if (!sketchTypes.includes(targetType)) {
-    die(`Invalid option --to ${options.to}; must be `);
+  if (targetType === 'folder') {
+    const targetDir = sketchPath.replace(/\.(html?|js)/i, '');
+    fs.mkdirSync(targetDir, { recursive: true });
+    sketch.files.forEach(file => {
+      const targetName = file === sketch.htmlFile ? 'index.html' : file;
+      fs.renameSync(path.join(sketch.dir, file), path.join(targetDir, targetName));
+      console.log(`Moved ${file} into new directory ${targetDir}`);
+    });
+    return;
   }
 
-  if (sketch.sketchType === options.to) {
+  if (sketch.structureType === targetType) {
     console.log('Nothing to do');
     return;
   }