+++ p5-analysis/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## Unreleased
+
+Changed:
+
+- New enum SketchStructureType
+
+Changed:
+
+- Sketch.sketchType -> Sketch.structureType; changed the enum values
+- It's only a single-directory sketch if the HTML file is named index.html
+
 ## [0.6.3] - 2021-10-05
 
 Fixed: