+++ p5-server/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## Unreleased
 
+- `p5 convert` is smarter, can convert into folders
+- It's only a single-directory sketch if the HTML file is named index.html
 - Rename Server#stop() -> Server#close()
 
 ## [0.8.2] - 2021-10-05